add more backup stories, sync by one provider

+++ packages/taler-wallet-webextension/clean_and_build.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
 # This file is in the public domain.
+[ "also-wallet" == "$1" ] && { pnpm -C ../taler-wallet-core/ compile || exit 1; }
 pnpm clean && pnpm compile && rm -rf extension/ && ./pack.sh  && (cd extension/ && unzip taler*.zip)
